+++ src/classes/RammerheadSessionFileCache.js
@@ -2,7 +2,7 @@ const fs = require('fs');
 const path = require('path');
 const RammerheadSessionAbstractStore = require('./RammerheadSessionAbstractStore');
 const RammerheadSession = require('./RammerheadSession');
-const logger = require('../util/logger');
+const RammerheadLogging = require('../classes/RammerheadLogging');
 
 // rh = rammerhead. extra f to distinguish between rhsession (folder) and rhfsession (file)
 const sessionFileExtension = '.rhfsession';
@@ -13,8 +13,10 @@ class RammerheadSessionFileCache extends RammerheadSessionAbstractStore {
      * @param {object} options
      * @param {string} options.saveDirectory - all cacheTimeouted sessions will be saved in this folder
      * to avoid storing all the sessions in the memory.
+     * @param {RammerheadLogging|undefined} options.logger
      * @param {number} options.cacheTimeout - timeout before saving cache to disk and deleting it from the cache
      * @param {number} options.cacheCheckInterval
+     * @param {boolean} options.deleteUnused - (default: true) if set to true, it deletes unused sessions when saving cache to disk
      * @param {object|null} options.staleCleanupOptions - set to null to disable cleaning up stale sessions
      * @param {number|null} options.staleCleanupOptions.staleTimeout - stale sessions that are inside saveDirectory that go over
      * this timeout will be deleted. Set to null to disable.
@@ -24,8 +26,10 @@ class RammerheadSessionFileCache extends RammerheadSessionAbstractStore {
      */
     constructor({
         saveDirectory = path.join(__dirname, '../../sessions'),
+        logger = new RammerheadLogging({ logLevel: 'disabled' }),
         cacheTimeout = 1000 * 60 * 20, // 20 minutes
-        cacheCheckInterval = 1000 * 60 * 10, // 10 minutes
+        cacheCheckInterval = 1000 * 60 * 10, // 10 minutes,
+        deleteUnused = true,
         staleCleanupOptions = {
             staleTimeout: 1000 * 60 * 60 * 24 * 1, // 1 day
             maxToLive: 1000 * 60 * 60 * 24 * 4, // four days
@@ -34,12 +38,16 @@ class RammerheadSessionFileCache extends RammerheadSessionAbstractStore {
     } = {}) {
         super();
         this.saveDirectory = saveDirectory;
+        this.logger = logger;
+        this.deleteUnused = deleteUnused;
+        this.cacheTimeout = cacheTimeout;
         /**
          * @type {Map.<string, RammerheadSession>}
          */
         this.cachedSessions = new Map();
-        setInterval(() => this._saveCacheToDisk(cacheTimeout), cacheCheckInterval).unref();
+        setInterval(() => this._saveCacheToDisk(), cacheCheckInterval).unref();
         if (staleCleanupOptions) {
+            this._removeStaleSessions(staleCleanupOptions.staleTimeout, staleCleanupOptions.maxToLive);
             setInterval(
                 () => this._removeStaleSessions(staleCleanupOptions.staleTimeout, staleCleanupOptions.maxToLive),
                 staleCleanupOptions.staleCheckInterval
@@ -65,26 +73,26 @@ class RammerheadSessionFileCache extends RammerheadSessionAbstractStore {
     }
     get(id, updateActiveTimestamp = true, cacheToMemory = true) {
         if (!this.has(id)) {
-            logger.debug(`(FileCache.get) ${id} does not exist`);
+            this.logger.debug(`(FileCache.get) ${id} does not exist`);
             return;
         }
 
-        logger.debug(`(FileCache.get) ${id}`);
+        this.logger.debug(`(FileCache.get) ${id}`);
         if (this.cachedSessions.has(id)) {
-            logger.debug(`(FileCache.get) returning memory cached session ${id}`);
+            this.logger.debug(`(FileCache.get) returning memory cached session ${id}`);
             return this.cachedSessions.get(id);
         }
 
         const session = RammerheadSession.DeserializeSession(id, fs.readFileSync(this._getSessionFilePath(id)));
 
         if (updateActiveTimestamp) {
-            logger.debug(`(FileCache.get) ${id} update active timestamp`);
+            this.logger.debug(`(FileCache.get) ${id} update active timestamp`);
             session.updateLastUsed();
         }
 
         if (cacheToMemory) {
             this.cachedSessions.set(id, session);
-            logger.debug(`(FileCache.get) saved ${id} into cache memory`);
+            this.logger.debug(`(FileCache.get) saved ${id} into cache memory`);
         }
 
         return session;
@@ -94,23 +102,23 @@ class RammerheadSessionFileCache extends RammerheadSessionAbstractStore {
 
         fs.writeFileSync(this._getSessionFilePath(id), new RammerheadSession().serializeSession());
 
-        logger.debug(`FileCache.add ${id}`);
+        this.logger.debug(`FileCache.add ${id}`);
 
         return this.get(id);
     }
     delete(id) {
-        logger.debug(`(FileCache.delete) deleting ${id}`);
+        this.logger.debug(`(FileCache.delete) deleting ${id}`);
         if (this.has(id)) {
             fs.unlinkSync(this._getSessionFilePath(id));
             this.cachedSessions.delete(id);
-            logger.debug(`(FileCache.delete) deleted ${id}`);
+            this.logger.debug(`(FileCache.delete) deleted ${id}`);
             return true;
         }
-        logger.debug(`(FileCache.delete) ${id} does not exist`);
+        this.logger.debug(`(FileCache.delete) ${id} does not exist`);
         return false;
     }
     close() {
-        logger.debug(`(FileCache.close) calling _saveCacheToDisk`);
+        this.logger.debug(`(FileCache.close) calling _saveCacheToDisk`);
         this._saveCacheToDisk(-1);
     }
 
@@ -130,42 +138,49 @@ class RammerheadSessionFileCache extends RammerheadSessionAbstractStore {
     _removeStaleSessions(staleTimeout, maxToLive) {
         const sessionIds = this.keysStore();
         let deleteCount = 0;
-        logger.debug(`(FileCache._removeStaleSessions) Need to go through ${sessionIds.length} sessions in store`);
+        this.logger.debug(`(FileCache._removeStaleSessions) Need to go through ${sessionIds.length} sessions in store`);
 
         const now = Date.now();
         for (const id of sessionIds) {
-            const session = this.get(id, false);
+            const session = this.get(id, false, false);
             if (
                 (staleTimeout && now - session.lastUsed > staleTimeout) ||
                 (maxToLive && now - session.createdAt > maxToLive)
             ) {
                 this.delete(id);
                 deleteCount++;
-                logger.debug(`(FileCache._removeStaleSessions) deleted ${id}`);
+                this.logger.debug(`(FileCache._removeStaleSessions) deleted ${id}`);
             }
         }
 
-        logger.debug(`(FileCache._removeStaleSessions) Deleted ${deleteCount} sessions from store`);
+        this.logger.debug(`(FileCache._removeStaleSessions) Deleted ${deleteCount} sessions from store`);
     }
     /**
      * @private
-     * @param {number} cacheTimeout
      */
-    _saveCacheToDisk(cacheTimeout) {
+    _saveCacheToDisk() {
         let deleteCount = 0;
-        logger.debug(`(FileCache._saveCacheToDisk) need to go through ${this.cachedSessions.size} sessions`);
+        this.logger.debug(`(FileCache._saveCacheToDisk) need to go through ${this.cachedSessions.size} sessions`);
 
         const now = Date.now();
         for (const [sessionId, session] of this.cachedSessions) {
-            if (now - session.lastUsed > cacheTimeout) {
-                fs.writeFileSync(this._getSessionFilePath(sessionId), session.serializeSession());
-                this.cachedSessions.delete(sessionId);
-                deleteCount++;
-                logger.debug(`(FileCache._saveCacheToDisk) removed ${sessionId} from memory and saved to store`);
+            if (now - session.lastUsed > this.cacheTimeout) {
+                if (session.lastUsed === session.createdAt && this.deleteUnused) {
+                    this.cachedSessions.delete(sessionId);
+                    deleteCount++;
+                    this.logger.debug(`(FileCache._saveCacheToDisk) deleted unused ${sessionId} from memory`);
+                } else {
+                    fs.writeFileSync(this._getSessionFilePath(sessionId), session.serializeSession());
+                    this.cachedSessions.delete(sessionId);
+                    deleteCount++;
+                    this.logger.debug(
+                        `(FileCache._saveCacheToDisk) removed ${sessionId} from memory and saved to store`
+                    );
+                }
             }
         }
 
-        logger.debug(`(FileCache._saveCacheToDisk) Removed ${deleteCount} sessions from memory`);
+        this.logger.debug(`(FileCache._saveCacheToDisk) Removed ${deleteCount} sessions from memory`);
     }
 }