+++ src/config.js
@@ -46,7 +46,12 @@ module.exports = {
         saveDirectory: path.join(__dirname, '../sessions'),
         cacheTimeout: 1000 * 60 * 20, // 20 minutes
         cacheCheckInterval: 1000 * 60 * 10, // 10 minutes
-        staleCleanupOptions: null
+        deleteUnused: true,
+        staleCleanupOptions: {
+            staleTimeout: 1000 * 60 * 60 * 24 * 3, // 3 days
+            maxToLive: null,
+            staleCheckInterval: 1000 * 60 * 60 * 6 // 6 hours
+        }
     },
 
     //// LOGGING CONFIGURATION ////