+++ src/index.js
@@ -1,7 +1,15 @@
 const RammerheadProxy = require('./classes/RammerheadProxy');
 const RammerheadLogging = require('./classes/RammerheadLogging');
 const RammerheadSession = require('./classes/RammerheadSession');
+const RammerheadSessionAbstractStore = require('./classes/RammerheadSessionAbstractStore');
+const RammerheadSessionFileCache = require('./classes/RammerheadSessionFileCache');
+const addStaticFilesToProxy = require('./util/addStaticDirToProxy');
 
 module.exports = {
-    RammerheadProxy, RammerheadLogging, RammerheadSession
+    RammerheadProxy,
+    RammerheadLogging,
+    RammerheadSession,
+    RammerheadSessionAbstractStore,
+    RammerheadSessionFileCache,
+    addStaticFilesToProxy
 };