+++ src/util/logger.js
@@ -1,7 +1,0 @@
-const RammerheadLogging = require('../classes/RammerheadLogging');
-const config = require('../config');
-
-module.exports = new RammerheadLogging({
-    logLevel: config.logLevel,
-    generatePrefix: config.generatePrefix
-});